+++ scripts/app.py
@@ -1,69 +1,243 @@
 import re
 import uuid
+import torch
 import os.path
 import chromadb
-import requests
 import tempfile
 import pandas as pd
 import gradio as gr
+from PIL import Image
 from re import Pattern
 from __init__ import *
-from llama_cpp import Llama
+from threading import Thread
 from datetime import datetime
 from gradio_modal import Modal
 from tinydb import TinyDB, where
+from functions.functions import *
 from collections import defaultdict
 from langchain.docstore.document import Document
 from typing import List, Optional, Tuple, Iterator
-from huggingface_hub.file_download import http_get
 from langchain_community.vectorstores import Chroma
-from langchain_community.embeddings import HuggingFaceEmbeddings
+from langchain_huggingface import HuggingFaceEmbeddings
 from langchain.text_splitter import RecursiveCharacterTextSplitter
-
+from transformers import MllamaForConditionalGeneration, AutoProcessor, TextIteratorStreamer
 
 os.environ["TOKENIZERS_PARALLELISM"] = "false"
 logger: logging.getLogger = get_logger(os.path.basename(__file__).replace(".py", "_") + str(datetime.now().date()))
 
 
-class LocalGPT:
+class SystemPromptManager:
     def __init__(self):
-        self.llm: Optional[Llama] = self.load_or_initialize_model()
+        self.mode = MODES[0]
+        self.system_prompt = ""
+
+    def set_system_prompt(self, system_prompt_input: str) -> None:
+        """
+        Setting prompt.
+        :param system_prompt_input: Prompt.
+        :return:
+        """
+        self.system_prompt = system_prompt_input
+
+    def set_current_mode(self, mode: str) -> gr.update:
+        """
+        Setting prompt.
+        :param mode: Mode.
+        :return:
+        """
+        self.mode = mode
+        self.set_system_prompt(self._get_default_system_prompt(mode))
+        # Update placeholder and allow interaction if default system prompt is set
+        if self.system_prompt:
+            return gr.update(placeholder=self.system_prompt, interactive=True)
+        # Update placeholder and disable interaction if no default system prompt is set
+        else:
+            return gr.update(placeholder=self.system_prompt, interactive=False)
+
+    @staticmethod
+    def _get_default_system_prompt(mode: str) -> str:
+        """
+        Returning prompt of mode.
+        :param mode: Mode.
+        :return: Prompt.
+        """
+        return QUERY_SYSTEM_PROMPT if mode == "DB" else LLM_SYSTEM_PROMPT
+
+
+class AnalyticsManager:
+    def __init__(self):
+        self.tiny_db: TinyDB = TinyDB(f'{QUESTIONS}/tiny_db.json', indent=4, ensure_ascii=False)
+
+    def get_analytics(self) -> pd.DataFrame:
+        """
+        Retrieves and returns analytics data from the database as a sorted DataFrame.
+
+        This method fetches all data entries from the `tiny_db` database, converts them into a DataFrame,
+        and sorts the records by the 'Старт обработки запроса' (Request Processing Start) column in
+        descending order if this column is present. If the column is missing, it returns the DataFrame unsorted.
+
+        :return: A DataFrame containing analytics data, optionally sorted by 'Старт обработки запроса'
+        in descending order.
+        """
+        try:
+            return pd.DataFrame(self.tiny_db.all()).sort_values('Старт обработки запроса', ascending=False)
+        except KeyError:
+            return pd.DataFrame(self.tiny_db.all())
+
+    def update_message_analytics(self, messages: List[dict], analyse=None):
+        """
+        Updates or inserts analytics data for the latest message in the database.
+
+        This function processes the last message in a list of message-answer pairs (`messages`). If the message
+        already exists in the database, it updates the stored answer, increments the repetition count, and
+        optionally adds a rating (`analyse`). If the message is new, it inserts a new record with the current
+        timestamp. Finally, it returns the updated analytics DataFrame.
+
+        :param messages: List of tuples where each tuple is a (message, answer) pair.
+        :param analyse: Optional; rating to assign to the message-answer pair. If not provided, defaults to None.
+        :return: A DataFrame containing the latest analytics data.
+        """
+        message = messages[-2]["content"] if messages else None
+        answer = messages[-1]["content"] if message else None
+        filter_query = where('Сообщения') == message
+        if result := self.tiny_db.search(filter_query):
+            if analyse is None:
+                self.tiny_db.update({
+                    'Ответы': answer,
+                    'Количество повторений': result[0]['Количество повторений'] + 1,
+                    'Старт обработки запроса': str(datetime.now())
+                }, cond=filter_query)
+            else:
+                self.tiny_db.update({'Оценка ответа': analyse}, cond=filter_query)
+                gr.Info("Отзыв ответу поставлен")
+        elif message is not None:
+            self.tiny_db.insert({
+                'Сообщения': message,
+                'Ответы': answer,
+                'Количество повторений': 1,
+                'Оценка ответа': None,
+                'Старт обработки запроса': str(datetime.now())
+            })
+        return self.get_analytics()
+
+
+class AuthManager:
+    def __init__(self, document_manager):
+        self.document_manager: DocumentManager = document_manager
+
+    @staticmethod
+    def login(username: str, password: str) -> dict:
+        """
+        Sends a login request to obtain an access token for the provided user credentials.
+
+        This function takes in a username and password, then sends a POST request to the
+        authentication endpoint to retrieve an access token. If authentication is successful,
+        the access token is returned along with a success flag. In case of failure, an error
+        message is logged, and the function returns failure information with an error message.
+
+        :param username: The username of the user attempting to authenticate.
+        :param password: The password of the user attempting to authenticate.
+        :return: A dictionary containing:
+                 - "access_token": The access token if authentication is successful, else None.
+                 - "is_success": Boolean indicating success (True) or failure (False).
+                 - "message": Error message if authentication fails, else not included.
+        """
+        try:
+            response = requests.post(
+                f"http://{IP_ADDRESS}/token",
+                data={"username": username, "password": password},
+                headers={"Content-Type": "application/x-www-form-urlencoded"},
+                timeout=10  # Adding a timeout
+            )
+            if response.status_code == 200:
+                return {"access_token": response.json().get("access_token"), "is_success": True}
+            error_detail = response.json().get("detail", "Unknown error")
+            logger.error(f"Login failed with status {response.status_code}: {error_detail}")
+            return {"access_token": None, "is_success": False, "message": error_detail}
+        except requests.RequestException as e:
+            logger.error(f"Error during login request: {e}")
+            return {"access_token": None, "is_success": False, "message": "Request failed"}
+
+    def update_user_ui_state(self, local_data: Optional[dict]):
+        """
+        Retrieves current user information and updates the user interface based on login status.
+
+        Validates the `local_data` to check if the user is authenticated. If authenticated, retrieves
+        user details from the server and prepares the UI to reflect the logged-in state. In case of failed
+        authentication or a logged-out state, it updates the UI to reflect this accordingly.
+
+        :param local_data: A dictionary with local user data, including an access token and login status.
+        :return: A list of UI updates for interface components based on user login status.
+        """
+        if isinstance(local_data, dict) and local_data.get("is_success", False):
+            response = requests.get(
+                f"http://{IP_ADDRESS}/users/me",
+                headers={"Authorization": f"Bearer {local_data['access_token']}"}
+            )
+            logger.info(f"User is {response.json().get('username')}")
+            is_logged_in = response.status_code == 200
+        else:
+            is_logged_in = False
+
+        obj_tabs = [local_data] + [gr.update(visible=is_logged_in) for _ in range(3)]
+        if is_logged_in:
+            obj_tabs.append(gr.update(value="Выйти", icon=LOGOUT_ICON))
+        else:
+            obj_tabs.append(gr.update(value="Войти", icon=LOGIN_ICON))
+        obj_tabs.append(gr.update(visible=not is_logged_in))
+        if isinstance(local_data, dict):
+            obj_tabs.append(local_data.get("message", MESSAGE_LOGIN))
+        else:
+            obj_tabs.append(MESSAGE_LOGIN)
+        obj_tabs.append(self.document_manager.list_ingested_documents())
+        return obj_tabs
+
+    def toggle_login_state(self, local_data: Optional[dict], login_btn: gr.component):
+        """
+        Handles user login/logout functionality and updates the UI accordingly.
+
+        This function checks the current user's login status using `local_data`. If the user is logged in,
+        it updates the UI to reflect a logged-out state, changing the button to "Login." If the user
+        is not logged in, it shows the login button and adjusts the visibility of other UI components.
+
+        :param local_data: A dictionary containing local user data, which may include an access token.
+        :param login_btn: The Gradio component representing the login button.
+        :return: A list of UI updates to be applied based on the user's login state.
+        """
+        data = self.update_user_ui_state(local_data)
+        is_logged_in = isinstance(data[0], dict) and data[0].get("access_token")
+
+        obj_tabs = [gr.update(visible=not is_logged_in)] + [gr.update(visible=False) for _ in range(3)]
+        obj_tabs.append(gr.update(value="Войти" if is_logged_in else login_btn))
+
+        return obj_tabs
+
+
+class DocumentManager:
+    def __init__(self):
         self.embeddings: HuggingFaceEmbeddings = HuggingFaceEmbeddings(
             model_name=EMBEDDER_NAME,
             cache_folder=MODELS_DIR
         )
         self.collection: str = "all-documents"
         self.db: Optional[Chroma] = None
-        self.mode: str = MODES[0]
-        self.system_prompt: str = self._get_default_system_prompt(self.mode)
-        self._queue = 0
-        self.tiny_db: TinyDB = TinyDB(f'{QUESTIONS}/tiny_db.json', indent=4, ensure_ascii=False)
 
-    @staticmethod
-    def load_or_initialize_model() -> Llama:
+    def initialize_database(self) -> Chroma:
         """
-        Loads and initializes the Llama model from a specified repository.
+        Loads the database connection using a persistent Chroma client.
 
-        This method creates necessary directories for model storage if they do not
-        already exist. It checks if the model file is present; if not, it downloads
-        the model from the specified repository. Finally, it initializes and returns
-        a Llama model instance with the specified configuration.
+        This method initializes a PersistentClient for Chroma, specifying the
+        path to the database directory. It then creates and returns a Chroma
+        instance that can be used to interact with the specified collection
+        using the defined embedding function.
 
-        :return: An instance of the Llama model initialized with the specified parameters.
+        :return: An instance of Chroma connected to the specified collection.
         """
-        os.makedirs(MODELS_DIR, exist_ok=True)
-        final_model_path: str = os.path.join(MODELS_DIR, MODEL_NAME)
-        os.makedirs(os.path.dirname(final_model_path), exist_ok=True)
-
-        if not os.path.exists(final_model_path):
-            with open(final_model_path, "wb") as f:
-                http_get(REPO, f)
-
-        return Llama(
-            n_gpu_layers=-1,
-            model_path=final_model_path,
-            n_ctx=CONTEXT_SIZE,
-            n_parts=1,
+        client = chromadb.PersistentClient(path=DB_DIR)
+        return Chroma(
+            client=client,
+            collection_name=self.collection,
+            embedding_function=self.embeddings,
         )
 
     @staticmethod
@@ -93,7 +267,7 @@ class LocalGPT:
             raise FileNotFoundError(f"Failed to load document at {file_path}") from e
 
     @staticmethod
-    def _process_text(text: str) -> str:
+    def _process_text(page_content: str) -> str:
         """
         Processes the input text by removing unnecessary lines and formatting it into a more readable form.
 
@@ -101,14 +275,14 @@ class LocalGPT:
         the remaining lines into a single string. If the resulting text is shorter
         than 10 characters, an empty string is returned.
 
-        :param text: The input string containing the text to be processed.
+        :param page_content: The input string containing the text to be processed.
         :return: A cleaned and formatted version of the input text. Returns an
                  empty string if the processed text is less than 10 characters long.
         """
-        lines: list = text.split("\n")
+        lines: list = page_content.split("\n")
         lines = [line for line in lines if len(line.strip()) > 2]
-        text = "\n".join(lines).strip()
-        return "" if len(text) < 10 else text
+        page_content = "\n".join(lines).strip()
+        return "" if len(page_content) < 10 else page_content
 
     def update_documents(self, fixed_documents: List[Document], ids: List[str]) -> tuple[bool, str]:
         """
@@ -196,220 +370,9 @@ class LocalGPT:
         os.chmod(FILES_DIR, 0o0777)
         return file_warning
 
-    @staticmethod
-    def _get_default_system_prompt(mode: str) -> str:
-        """
-        Returning prompt of mode.
-        :param mode: Mode.
-        :return: Prompt.
-        """
-        return QUERY_SYSTEM_PROMPT if mode == "DB" else LLM_SYSTEM_PROMPT
-
-    def _set_system_prompt(self, system_prompt_input: str) -> None:
-        """
-        Setting prompt.
-        :param system_prompt_input: Prompt.
-        :return:
-        """
-        self.system_prompt = system_prompt_input
-
-    def _set_current_mode(self, mode: str) -> gr.update:
-        """
-        Setting prompt.
-        :param mode: Mode.
-        :return:
-        """
-        self.mode = mode
-        self._set_system_prompt(self._get_default_system_prompt(mode))
-        # Update placeholder and allow interaction if default system prompt is set
-        if self.system_prompt:
-            return gr.update(placeholder=self.system_prompt, interactive=True)
-        # Update placeholder and disable interaction if no default system prompt is set
-        else:
-            return gr.update(placeholder=self.system_prompt, interactive=False)
-
-    def generate_chat_completion(
-        self,
-        history: List[list],
-        retrieved_docs: str,
-        mode: str,
-        top_k: int,
-        top_p: float,
-        temp: float,
-        uid: str
-    ):
-        """
-        Generate a chat completion response based on user history and retrieved documents.
-
-        This function constructs a message prompt for the language model, incorporating user history,
-        context from retrieved documents, and various parameters that influence the response generation.
-
-        :param history: A list of tuples containing the conversation history, where each tuple consists of
-                        the user's message and the assistant's response.
-        :param retrieved_docs: A string containing documents retrieved for context, which may include
-                               HTML anchor tags.
-        :param mode: A string indicating the mode of response generation.
-        :param top_k: An integer specifying the number of top tokens to sample from during generation.
-        :param top_p: A float representing the cumulative probability threshold for token sampling (nucleus sampling).
-        :param temp: A float that controls the randomness of the output; higher values produce more
-                     random responses.
-        :param uid: A unique identifier for the user, used for logging purposes.
-        :return: A generator for the chat completion response and a list of files extracted from the
-                 retrieved documents.
-        """
-        last_user_message: str = history[-1][0]
-        files = re.findall(r'<a\s+[^>]*>(.*?)</a>', retrieved_docs)
-        for file in files:
-            retrieved_docs = re.sub(fr'<a\s+[^>]*>{file}</a>', file, retrieved_docs)
-        if retrieved_docs and mode == MODES[0]:
-            last_user_message = (
-                f"Контекст: {retrieved_docs}\n\nИспользуя только контекст, ответь на вопрос: "
-                f"{last_user_message}"
-            )
-        logger.info(f"The question has been fully formed [uid - {uid}]")
-        dialog_history: List[dict] = []
-        for user_msg, bot_msg in history[-4:-1]:
-            dialog_history.extend((
-                {"role": "user", "content": user_msg},
-                {"role": "assistant", "content": bot_msg}
-            ))
-        generator = self.llm.create_chat_completion(
-            messages=[
-                {"role": "system", "content": self.system_prompt},
-                *dialog_history,
-                {"role": "user", "content": last_user_message},
-            ],
-            stream=True,
-            temperature=temp,
-            top_k=top_k,
-            top_p=top_p
-        )
-        return generator, files
-
-    @staticmethod
-    def _add_source_references(
-        history: List[List[str]],
-        scores: List[float],
-        files: List[str],
-        partial_text: str,
-        threshold: float = 0.44
-    ) -> List[List[str]]:
-        """
-        Appends file source references to the final response text based on score thresholds and
-        updates conversation history.
-        This method adds a list of file references to the response text if files are provided, adjusting based on score
-        threshold conditions. The updated text is appended to the most recent assistant message
-        in the conversation history.
-        :param history: List representing conversation history as pairs of user and assistant messages.
-        :param scores: List of floats representing confidence scores associated with each file, determining if
-                       file sources should be appended.
-        :param files: List of file names or identifiers to include as sources in the response.
-        :param partial_text: The assistant's partial response text to which sources will be appended if files exist.
-        :param threshold: The score threshold to determine whether all sources are appended or only the top source.
-        :return: Updated conversation history with appended source information if conditions are met.
-        """
-        if files:
-            partial_text += SOURCES_SEPARATOR
-            sources_text = [f"{index}. {source}" for index, source in enumerate(files, start=1)]
-            if scores and scores[0] < threshold:
-                partial_text += "\n\n\n".join(sources_text)
-            elif scores:
-                partial_text += sources_text[0]
-            history[-1][1] = partial_text
-        return history
-
-    def generate_response_stream(
-        self,
-        history: List[List[str]],
-        mode: str,
-        retrieved_docs: str,
-        top_p: float,
-        top_k: int,
-        temp: float,
-        scores: List[float],
-        uid: str
-    ) -> Iterator[List[List[str]]]:
-        """
-        Generates a response based on user input, context, and retrieved documents.
-
-        This function orchestrates the process of generating a response by interacting
-        with the chat completion generator.
-        The response is built incrementally, updating the conversation history as tokens are received.
-        In case of an error (e.g., due to large context size), an error message is appended to the response.
-        Once complete, source file references are appended to the final output.
-
-        :param history: List of conversation pairs (user input and bot responses).
-        :param mode: Operation mode, which influences the use of context in responses.
-        :param retrieved_docs: Relevant documents retrieved to help answer the user query.
-        :param top_p: Probability for nucleus sampling, controls randomness of responses.
-        :param top_k: Number of top probable tokens considered for generation.
-        :param temp: Temperature for response generation, influencing creativity.
-        :param scores: List of scores associated with retrieved documents for source filtering.
-        :param uid: Unique identifier for the current user session, useful for logging and tracking.
-        :return: Yields updated conversation history after each token generated, and the final response with sources.
-        """
-        logger.info(f"Preparing to generate a response based on context and history [uid - {uid}]")
-        if not history or not history[-1][0]:
-            yield history[:-1]
-            return
-
-        partial_text = ""
-        logger.info(f"Beginning response generation [uid - {uid}]")
-
-        generator, files = self.generate_chat_completion(
-            history=history,
-            retrieved_docs=retrieved_docs,
-            mode=mode,
-            top_k=top_k,
-            top_p=top_p,
-            temp=temp,
-            uid=uid
-        )
-
-        try:
-            for token in generator:
-                for data in token["choices"]:  # type: ignore
-                    partial_text += data["delta"].get("content", "")  # type: ignore
-                    history[-1][1] = partial_text
-                    yield history
-        except Exception as ex:
-            logger.error(f"Error - {ex}")
-            partial_text += (
-                "\nContext too large. Try reducing it or adjusting the output context amount in settings."
-            )
-            history[-1][1] = partial_text
-            yield history
-
-        logger.info(f"Response generation completed [uid - {uid}]")
-        yield self._add_source_references(history, scores, files, partial_text)
-        self._queue -= 1
-
-    def add_user_message(self, message: str, history: Optional[List]):
-        """
-        Adds a new user message to the conversation history and generates a unique session identifier.
-
-        This function appends the user's input message to the conversation history and increments the
-        queue counter to indicate a pending response generation. If history is not provided, a new
-        conversation history is initialized.
-
-        :param message: The user's input message to be added to the conversation history.
-        :param history: The existing conversation history as a list of message pairs (user, bot responses).
-                        Each pair is a list, with the second item initially set to None for new user messages.
-        :return: A tuple containing an empty string (for response text), the updated conversation history,
-                 and a unique session ID (uid).
-        """
-        uid = uuid.uuid4()
-        logger.info(f"Processing the question. Queue - {self._queue}. UID - [{uid}]")
-        if history is None:
-            history = []
-        new_history = history + [[message, None]]
-        self._queue += 1
-        logger.info(f"The question has been processed. UID - [{uid}]")
-        return "", new_history, uid
-
     def retrieve_documents(
         self,
-        history: List[List[Optional[str]]],
+        history: List[dict],
         collection_radio: str,
         k_documents: int,
         uid: str
@@ -429,8 +392,13 @@ class LocalGPT:
         :param uid: The unique identifier for the current session, used for logging.
         :return: A tuple with a formatted string of retrieved documents and a list of their similarity scores.
         """
-        if not self.db or collection_radio != MODES[0] or not history or not history[-1][0]:
-            return "Documents will appear after questions are asked", []
+        if (
+            not self.db
+            or collection_radio != MODES[0]
+            or not history
+            or history[-1]["role"] != "user"
+        ):
+            return "Появятся после задавания вопросов", []
 
         last_user_message = history[-1][0]
         docs = self.db.similarity_search_with_score(last_user_message, k_documents)
@@ -447,7 +415,7 @@ class LocalGPT:
             scores.append(score)
             data[document] += f"\n\nScore: {score}, Text: {doc[0].page_content}"
 
-        list_data: list = [f"{doc}\n\n{text}" for doc, text in data.items()]
+        list_data: list = [f"{doc}\n\n{page_content}" for doc, page_content in data.items()]
         logger.info(f"Retrieved context from database for collection '{collection_radio}' [uid - {uid}]")
 
         if not list_data:
@@ -455,7 +423,7 @@ class LocalGPT:
 
         return "\n\n\n".join(list_data), scores
 
-    def _list_ingested_documents(self):
+    def list_ingested_documents(self):
         """
         Loads the database and retrieves a list of ingested document filenames.
 
@@ -491,167 +459,246 @@ class LocalGPT:
                 if os.path.basename(ingested_document["source"]) in documents
             ]:
                 self.db.delete(for_delete_ids)
-            return self._list_ingested_documents()
+            return self.list_ingested_documents()
         except Exception as e:
             logger.error(f"Error during document deletion: {e}")
             return gr.update(choices=[])
 
-    def get_analytics(self) -> pd.DataFrame:
-        """
-        Retrieves and returns analytics data from the database as a sorted DataFrame.
 
-        This method fetches all data entries from the `tiny_db` database, converts them into a DataFrame,
-        and sorts the records by the 'Старт обработки запроса' (Request Processing Start) column in
-        descending order if this column is present. If the column is missing, it returns the DataFrame unsorted.
-
-        :return: A DataFrame containing analytics data, optionally sorted by 'Старт обработки запроса'
-        in descending order.
-        """
-        try:
-            return pd.DataFrame(self.tiny_db.all()).sort_values('Старт обработки запроса', ascending=False)
-        except KeyError:
-            return pd.DataFrame(self.tiny_db.all())
-
-    def _update_message_analytics(self, messages: List[List], analyse=None):
-        """
-        Updates or inserts analytics data for the latest message in the database.
-
-        This function processes the last message in a list of message-answer pairs (`messages`). If the message
-        already exists in the database, it updates the stored answer, increments the repetition count, and
-        optionally adds a rating (`analyse`). If the message is new, it inserts a new record with the current
-        timestamp. Finally, it returns the updated analytics DataFrame.
-
-        :param messages: List of tuples where each tuple is a (message, answer) pair.
-        :param analyse: Optional; rating to assign to the message-answer pair. If not provided, defaults to None.
-        :return: A DataFrame containing the latest analytics data.
-        """
-        message = messages[-1][0] if messages else None
-        answer = messages[-1][1] if message else None
-        filter_query = where('Сообщения') == message
-        if result := self.tiny_db.search(filter_query):
-            if analyse is None:
-                self.tiny_db.update({
-                    'Ответы': answer,
-                    'Количество повторений': result[0]['Количество повторений'] + 1,
-                    'Старт обработки запроса': str(datetime.now())
-                }, cond=filter_query)
-            else:
-                self.tiny_db.update({'Оценка ответа': analyse}, cond=filter_query)
-                gr.Info("Отзыв ответу поставлен")
-        elif message is not None:
-            self.tiny_db.insert({
-                'Сообщения': message,
-                'Ответы': answer,
-                'Количество повторений': 1,
-                'Оценка ответа': None,
-                'Старт обработки запроса': str(datetime.now())
-            })
-        return self.get_analytics()
+class LocalGPT:
+    def __init__(self):
+        self.model, self.processor = self.load_or_initialize_model()
+        self._queue: int = 0
+        self.document_manager: DocumentManager = DocumentManager()
+        self.analytics_manager: AnalyticsManager = AnalyticsManager()
+        self.auth_manager: AuthManager = AuthManager(self.document_manager)
+        self.prompt_manager: SystemPromptManager = SystemPromptManager()
 
     @staticmethod
-    def login(username: str, password: str) -> dict:
+    def load_or_initialize_model():
         """
-        Sends a login request to obtain an access token for the provided user credentials.
+        Loads and initializes the Llama model from a specified repository.
 
-        This function takes in a username and password, then sends a POST request to the
-        authentication endpoint to retrieve an access token. If authentication is successful,
-        the access token is returned along with a success flag. In case of failure, an error
-        message is logged, and the function returns failure information with an error message.
+        This method creates necessary directories for model storage if they do not
+        already exist. It checks if the model file is present; if not, it downloads
+        the model from the specified repository. Finally, it initializes and returns
+        a Llama model instance with the specified configuration.
 
-        :param username: The username of the user attempting to authenticate.
-        :param password: The password of the user attempting to authenticate.
-        :return: A dictionary containing:
-                 - "access_token": The access token if authentication is successful, else None.
-                 - "is_success": Boolean indicating success (True) or failure (False).
-                 - "message": Error message if authentication fails, else not included.
+        :return: An instance of the Llama model initialized with the specified parameters.
         """
-        try:
-            response = requests.post(
-                f"http://{IP_ADDRESS}/token",
-                data={"username": username, "password": password},
-                headers={"Content-Type": "application/x-www-form-urlencoded"},
-                timeout=10  # Adding a timeout
-            )
-            if response.status_code == 200:
-                return {"access_token": response.json().get("access_token"), "is_success": True}
-            error_detail = response.json().get("detail", "Unknown error")
-            logger.error(f"Login failed with status {response.status_code}: {error_detail}")
-            return {"access_token": None, "is_success": False, "message": error_detail}
-        except requests.RequestException as e:
-            logger.error(f"Error during login request: {e}")
-            return {"access_token": None, "is_success": False, "message": "Request failed"}
-
-    def update_user_ui_state(self, local_data: Optional[dict]):
-        """
-        Retrieves current user information and updates the user interface based on login status.
-
-        Validates the `local_data` to check if the user is authenticated. If authenticated, retrieves
-        user details from the server and prepares the UI to reflect the logged-in state. In case of failed
-        authentication or a logged-out state, it updates the UI to reflect this accordingly.
-
-        :param local_data: A dictionary with local user data, including an access token and login status.
-        :return: A list of UI updates for interface components based on user login status.
-        """
-        if isinstance(local_data, dict) and local_data.get("is_success", False):
-            response = requests.get(
-                f"http://{IP_ADDRESS}/users/me",
-                headers={"Authorization": f"Bearer {local_data['access_token']}"}
-            )
-            logger.info(f"User is {response.json().get('username')}")
-            is_logged_in = response.status_code == 200
-        else:
-            is_logged_in = False
-
-        obj_tabs = [local_data] + [gr.update(visible=is_logged_in) for _ in range(3)]
-        if is_logged_in:
-            obj_tabs.append(gr.update(value="Выйти", icon=LOGOUT_ICON))
-        else:
-            obj_tabs.append(gr.update(value="Войти", icon=LOGIN_ICON))
-        obj_tabs.append(gr.update(visible=not is_logged_in))
-        if isinstance(local_data, dict):
-            obj_tabs.append(local_data.get("message", MESSAGE_LOGIN))
-        else:
-            obj_tabs.append(MESSAGE_LOGIN)
-        obj_tabs.append(self._list_ingested_documents())
-        return obj_tabs
-
-    def _toggle_login_state(self, local_data: Optional[dict], login_btn: gr.component):
-        """
-        Handles user login/logout functionality and updates the UI accordingly.
-
-        This function checks the current user's login status using `local_data`. If the user is logged in,
-        it updates the UI to reflect a logged-out state, changing the button to "Login." If the user
-        is not logged in, it shows the login button and adjusts the visibility of other UI components.
-
-        :param local_data: A dictionary containing local user data, which may include an access token.
-        :param login_btn: The Gradio component representing the login button.
-        :return: A list of UI updates to be applied based on the user's login state.
-        """
-        data = self.update_user_ui_state(local_data)
-        is_logged_in = isinstance(data[0], dict) and data[0].get("access_token")
-
-        obj_tabs = [gr.update(visible=not is_logged_in)] + [gr.update(visible=False) for _ in range(3)]
-        obj_tabs.append(gr.update(value="Войти" if is_logged_in else login_btn))
-
-        return obj_tabs
-
-    def initialize_database(self) -> Chroma:
-        """
-        Loads the database connection using a persistent Chroma client.
-
-        This method initializes a PersistentClient for Chroma, specifying the
-        path to the database directory. It then creates and returns a Chroma
-        instance that can be used to interact with the specified collection
-        using the defined embedding function.
-
-        :return: An instance of Chroma connected to the specified collection.
-        """
-        client = chromadb.PersistentClient(path=DB_DIR)
-        return Chroma(
-            client=client,
-            collection_name=self.collection,
-            embedding_function=self.embeddings,
+        model = MllamaForConditionalGeneration.from_pretrained(
+            REPO_ID,
+            torch_dtype=torch.bfloat16,
+            cache_dir=MODELS_DIR
         )
+        processor = AutoProcessor.from_pretrained(REPO_ID)
+        return model, processor
+
+    def generate_chat_completion(
+        self,
+        history: List[dict],
+        retrieved_docs: str,
+        mode: str,
+        uid: str
+    ):
+        """
+        Generate a chat completion response based on user history and retrieved documents.
+
+        This function constructs a message prompt for the language model, incorporating user history,
+        context from retrieved documents, and various parameters that influence the response generation.
+
+        :param history: A list of tuples containing the conversation history, where each tuple consists of
+                        the user's message and the assistant's response.
+        :param retrieved_docs: A string containing documents retrieved for context, which may include
+                               HTML anchor tags.
+        :param mode: A string indicating the mode of response generation.
+        :param uid: A unique identifier for the user, used for logging purposes.
+        :return: A generator for the chat completion response and a list of files extracted from the
+                 retrieved documents.
+        """
+        last_user_message: str = history[-1].get("content")
+        files = re.findall(r'<a\s+[^>]*>(.*?)</a>', retrieved_docs)
+        for file in files:
+            retrieved_docs = re.sub(fr'<a\s+[^>]*>{file}</a>', file, retrieved_docs)
+        if retrieved_docs and mode == MODES[0]:
+            last_user_message = (
+                f"Контекст: {retrieved_docs}\n\nИспользуя только контекст, ответь на вопрос: "
+                f"{last_user_message}"
+            )
+        logger.info(f"The question has been fully formed [uid - {uid}]")
+
+        images = []
+        pair_count = 0  # Счетчик пар "user-assistant"
+        temp_history = []  # Временный список для обратного добавления диалогов
+        for message in reversed(history):
+            if message["role"] == "user" and pair_count == 0:
+                continue
+            if message["role"] == "user" and isinstance(message["content"], tuple):
+                images.append(Image.open(message["content"][0]).convert("RGB"))
+                temp_history.append({
+                    "role": "user",
+                    "content": [
+                        {"type": "text", "text": temp_history[-1]["content"]},
+                        {"type": "image"}
+                    ]
+                })
+            elif message["role"] == "user":
+                temp_history.append({
+                    "role": "user",
+                    "content": [{"type": "text", "text": message["content"]}],
+                })
+            if message["role"] == "assistant":
+                pair_count += 1
+                temp_history.append({
+                    "role": "assistant",
+                    "content": [
+                        {"type": "text", "text": message["content"]}
+                    ],
+                })
+            if pair_count == 3:
+                break
+
+        dialog_history: List[dict] = list(reversed(temp_history))
+        messages = [
+            *dialog_history
+        ]
+        if len(history) > 1 and isinstance(history[-2].get("content"), tuple):
+            images.append(Image.open(history[-2].get("content")[0]).convert("RGB"))
+            messages.append({
+                "role": "user",
+                "content": [
+                    {"type": "text", "text": last_user_message},
+                    {"type": "image"}
+                ]
+            })
+        else:
+            messages.append({
+                "role": "user",
+                "content": [
+                    {"type": "text", "text": last_user_message}
+                ]
+            })
+
+        texts = self.processor.apply_chat_template(messages, add_generation_prompt=True)
+        if not images:
+            inputs = self.processor(text=texts, return_tensors="pt")
+        else:
+            inputs = self.processor(text=texts, images=images, return_tensors="pt")
+
+        streamer = TextIteratorStreamer(self.processor, skip_special_tokens=True, skip_prompt=True)
+        generation_kwargs = dict(inputs, streamer=streamer, max_new_tokens=MAX_NEW_TOKENS)
+        return generation_kwargs, streamer, files
+
+    @staticmethod
+    def _add_source_references(
+        history: List[dict],
+        scores: List[float],
+        files: List[str],
+        partial_text: str,
+        threshold: float = 0.44
+    ) -> List[dict]:
+        """
+        Appends file source references to the final response text based on score thresholds and
+        updates conversation history.
+        This method adds a list of file references to the response text if files are provided, adjusting based on score
+        threshold conditions. The updated text is appended to the most recent assistant message
+        in the conversation history.
+        :param history: List representing conversation history as pairs of user and assistant messages.
+        :param scores: List of floats representing confidence scores associated with each file, determining if
+                       file sources should be appended.
+        :param files: List of file names or identifiers to include as sources in the response.
+        :param partial_text: The assistant's partial response text to which sources will be appended if files exist.
+        :param threshold: The score threshold to determine whether all sources are appended or only the top source.
+        :return: Updated conversation history with appended source information if conditions are met.
+        """
+        if files:
+            partial_text += SOURCES_SEPARATOR
+            sources_text = [f"{index}. {source}" for index, source in enumerate(files, start=1)]
+            if scores and scores[0] < threshold:
+                partial_text += "\n\n\n".join(sources_text)
+            elif scores:
+                partial_text += sources_text[0]
+            history[-1]["content"] = partial_text
+        return history
+
+    def generate_response_stream(
+        self,
+        history: List[dict],
+        mode: str,
+        retrieved_docs: str,
+        scores: List[float],
+        uid: str
+    ) -> Iterator[List[List[str]]]:
+        """
+        Generates a response based on user input, context, and retrieved documents.
+
+        This function orchestrates the process of generating a response by interacting
+        with the chat completion generator.
+        The response is built incrementally, updating the conversation history as tokens are received.
+        In case of an error (e.g., due to large context size), an error message is appended to the response.
+        Once complete, source file references are appended to the final output.
+
+        :param history: List of conversation pairs (user input and bot responses).
+        :param mode: Operation mode, which influences the use of context in responses.
+        :param retrieved_docs: Relevant documents retrieved to help answer the user query.
+        :param scores: List of scores associated with retrieved documents for source filtering.
+        :param uid: Unique identifier for the current user session, useful for logging and tracking.
+        :return: Yields updated conversation history after each token generated, and the final response with sources.
+        """
+        logger.info(f"Preparing to generate a response based on context and history [uid - {uid}]")
+        if not history or not history[-1].get("role"):
+            yield history[:-1]
+            return
+
+        partial_text = ""
+        logger.info(f"Beginning response generation [uid - {uid}]")
+
+        generation_kwargs, streamer, files = self.generate_chat_completion(
+            history=history,
+            retrieved_docs=retrieved_docs,
+            mode=mode,
+            uid=uid
+        )
+
+        history.append({"role": "assistant", "content": None})
+        thread = Thread(target=self.model.generate, kwargs=generation_kwargs)
+        thread.start()
+        buffer = ""
+
+        for new_text in streamer:
+            buffer += new_text
+            history[-1]["content"] = buffer
+            yield history
+
+        logger.info(f"Response generation completed [uid - {uid}]")
+        yield self._add_source_references(history, scores, files, partial_text)
+        self._queue -= 1
+        _ = self.analytics_manager.update_message_analytics(history)
+
+    def add_user_message(self, message: dict, history: Optional[List]):
+        """
+        Adds a new user message to the conversation history and generates a unique session identifier.
+
+        This function appends the user's input message to the conversation history and increments the
+        queue counter to indicate a pending response generation. If history is not provided, a new
+        conversation history is initialized.
+
+        :param message: The user's input message to be added to the conversation history.
+        :param history: The existing conversation history as a list of message pairs (user, bot responses).
+                        Each pair is a list, with the second item initially set to None for new user messages.
+        :return: A tuple containing an empty string (for response text), the updated conversation history,
+                 and a unique session ID (uid).
+        """
+        uid = uuid.uuid4()
+        logger.info(f"Processing the question. Queue - {self._queue}. UID - [{uid}]")
+        if history is None:
+            history = []
+        if message["files"]:
+            history.append({"role": "user", "content": message["files"]})
+        history.append({"role": "user", "content": message["text"]})
+        self._queue += 1
+        logger.info(f"The question has been processed. UID - [{uid}]")
+        return "", history, uid
 
     def launch_ui(self):
         """
@@ -672,21 +719,7 @@ class LocalGPT:
         """
         with gr.Blocks(
             title="LocalGPT",
-            theme=gr.themes.Soft().set(
-                body_background_fill="white",
-                block_background_fill="#e1e5e8",
-                block_label_background_fill="#2042b9",
-                block_label_background_fill_dark="#2042b9",
-                block_label_text_color="white",
-                checkbox_label_background_fill_selected="#1f419b",
-                checkbox_label_background_fill_selected_dark="#1f419b",
-                checkbox_background_color_selected="#111d3d",
-                checkbox_background_color_selected_dark="#111d3d",
-                input_background_fill="#e1e5e8",
-                button_primary_background_fill="#1f419b",
-                button_primary_background_fill_dark="#1f419b",
-                shadow_drop_lg="5px 5px 5px 5px rgb(0 0 0 / 0.1)"
-            ),
+            theme=gr.themes.ocean.Ocean(),
             css=BLOCK_CSS
         ) as demo:
             # Ваш логотип и текст заголовка
@@ -705,17 +738,19 @@ class LocalGPT:
 
             with gr.Tab("Чат"):
                 with gr.Row():
-                    collection_radio = gr.Radio(
-                        choices=MODES,
-                        value=self.mode,
-                        show_label=False
-                    )
+                    with gr.Column():
+                        collection_radio = gr.Radio(
+                            choices=MODES,
+                            value=self.prompt_manager.mode,
+                            show_label=False
+                        )
 
                 with gr.Row():
                     with gr.Column(scale=10):
                         chatbot = gr.Chatbot(
-                            label="Диалог",
+                            label=f"LLM: {REPO_ID}",
                             height=500,
+                            type="messages",
                             show_copy_button=True,
                             show_share_button=True,
                             avatar_images=(
@@ -726,18 +761,16 @@ class LocalGPT:
 
                 with gr.Row():
                     with gr.Column(scale=20):
-                        msg = gr.Textbox(
+                        msg = gr.MultimodalTextbox(
                             label="Отправить сообщение",
-                            show_label=False,
                             placeholder="👉 Напишите запрос",
-                            container=False
+                            show_label=False
                         )
-                    with gr.Column(scale=3, min_width=100):
-                        submit = gr.Button("📤 Отправить", variant="primary")
 
                 with gr.Row(elem_id="buttons"):
                     like = gr.Button(value="👍 Понравилось")
                     dislike = gr.Button(value="👎 Не понравилось")
+                    stop_btn = gr.Button(value="🛑 Остановить")
                     clear = gr.Button(value="🗑️ Очистить")
 
                 with gr.Row():
@@ -759,14 +792,14 @@ class LocalGPT:
                         files_selected = gr.Dropdown(
                             choices=None,
                             label="Выберите файлы для удаления",
-                            value=None,
+                            value="",
                             multiselect=True
                         )
                         delete = gr.Button("🧹 Удалить", variant="primary")
 
             with gr.Tab("Настройки", visible=False) as settings_tab:
                 with gr.Row(elem_id="model_selector_row"):
-                    models = [MODEL_NAME]
+                    models = [REPO_ID]
                     gr.Dropdown(
                         choices=models,
                         value=models[0],
@@ -801,31 +834,6 @@ class LocalGPT:
                             interactive=True,
                             label="Пересечение"
                         )
-                    with gr.Tab(label="Параметры генерации"):
-                        top_p = gr.Slider(
-                            minimum=0.0,
-                            maximum=1.0,
-                            value=0.9,
-                            step=0.05,
-                            interactive=True,
-                            label="Top-p",
-                        )
-                        top_k = gr.Slider(
-                            minimum=10,
-                            maximum=100,
-                            value=80,
-                            step=5,
-                            interactive=True,
-                            label="Top-k",
-                        )
-                        temp = gr.Slider(
-                            minimum=0.0,
-                            maximum=2.0,
-                            value=0.1,
-                            step=0.1,
-                            interactive=True,
-                            label="Temp"
-                        )
 
                 with gr.Accordion("Системный промпт", open=False):
                     system_prompt = gr.Textbox(
@@ -835,7 +843,7 @@ class LocalGPT:
                     )
                     # On blur, set system prompt to use in queries
                     system_prompt.blur(
-                        self._set_system_prompt,
+                        self.prompt_manager.set_system_prompt,
                         inputs=system_prompt,
                     )
 
@@ -851,7 +859,7 @@ class LocalGPT:
                 with gr.Row():
                     with gr.Column():
                         analytics = gr.DataFrame(
-                            value=self.get_analytics,  # type: ignore
+                            value=self.analytics_manager.get_analytics,  # type: ignore
                             interactive=False,
                             wrap=True
                         )
@@ -876,11 +884,11 @@ class LocalGPT:
                     cancel_login = gr.Button("⛔ Отмена", variant="secondary")
 
             submit_login.click(
-                fn=self.login,
+                fn=self.auth_manager.login,
                 inputs=[login, password],
                 outputs=[local_data]
             ).success(
-                fn=self.update_user_ui_state,
+                fn=self.auth_manager.update_user_ui_state,
                 inputs=[local_data],
                 outputs=[local_data, documents_tab, settings_tab, logging_tab, login_btn, modal, message_login]
             ).success(
@@ -891,7 +899,7 @@ class LocalGPT:
             )
 
             login_btn.click(
-                fn=self._toggle_login_state,
+                fn=self.auth_manager.toggle_login_state,
                 inputs=[local_data, login_btn],
                 outputs=[modal, documents_tab, settings_tab, logging_tab, login_btn]
             ).success(
@@ -907,78 +915,50 @@ class LocalGPT:
             )
 
             collection_radio.change(
-                fn=self._set_current_mode,
+                fn=self.prompt_manager.set_current_mode,
                 inputs=collection_radio,
                 outputs=system_prompt
             )
 
             # Upload files
             upload_files.upload(
-                fn=self.index_documents,
+                fn=self.document_manager.index_documents,
                 inputs=[upload_files, chunk_size, chunk_overlap],
                 outputs=[file_warning],
                 queue=True
             ).success(
-                fn=self._list_ingested_documents,
+                fn=self.document_manager.list_ingested_documents,
                 outputs=files_selected
             )
 
             # Delete documents from db
             delete.click(
-                fn=self.delete_documents,
+                fn=self.document_manager.delete_documents,
                 inputs=files_selected,
                 outputs=[files_selected]
             )
 
             # Pressing Enter
-            msg.submit(
+            click_msg_event = msg.submit(
                 fn=self.add_user_message,
                 inputs=[msg, chatbot],
                 outputs=[msg, chatbot, uid],
                 queue=False,
             ).success(
-                fn=self.retrieve_documents,
+                fn=self.document_manager.retrieve_documents,
                 inputs=[chatbot, collection_radio, k_documents, uid],
                 outputs=[retrieved_docs, scores],
                 queue=True,
             ).success(
                 fn=self.generate_response_stream,
-                inputs=[chatbot, collection_radio, retrieved_docs, top_p, top_k, temp, scores, uid],
+                inputs=[chatbot, collection_radio, retrieved_docs, scores, uid],
                 outputs=chatbot,
                 queue=True
-            ).success(
-                fn=self._update_message_analytics,
-                inputs=chatbot,
-                outputs=analytics,
-                queue=True,
-            )
-
-            # Pressing the button
-            submit.click(
-                fn=self.add_user_message,
-                inputs=[msg, chatbot],
-                outputs=[msg, chatbot, uid],
-                queue=False,
-            ).success(
-                fn=self.retrieve_documents,
-                inputs=[chatbot, collection_radio, k_documents, uid],
-                outputs=[retrieved_docs, scores],
-                queue=True,
-            ).success(
-                fn=self.generate_response_stream,
-                inputs=[chatbot, collection_radio, retrieved_docs, top_p, top_k, temp, scores, uid],
-                outputs=chatbot,
-                queue=True
-            ).success(
-                fn=self._update_message_analytics,
-                inputs=chatbot,
-                outputs=analytics,
-                queue=True,
             )
 
             # Like
             like.click(
-                fn=self._update_message_analytics,
+                fn=self.analytics_manager.update_message_analytics,
                 inputs=[chatbot, like],
                 outputs=[analytics],
                 queue=True,
@@ -986,7 +966,7 @@ class LocalGPT:
 
             # Dislike
             dislike.click(
-                fn=self._update_message_analytics,
+                fn=self.analytics_manager.update_message_analytics,
                 inputs=[chatbot, dislike],
                 outputs=[analytics],
                 queue=True,
@@ -1001,8 +981,16 @@ class LocalGPT:
                 js=JS
             )
 
+            # Stop generation
+            stop_btn.click(
+                fn=None,
+                inputs=None,
+                outputs=None,
+                cancels=[click_msg_event]
+            )
+
             demo.load(
-                fn=self.update_user_ui_state,
+                fn=self.auth_manager.update_user_ui_state,
                 inputs=[local_data],
                 outputs=[
                     local_data,
